# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "central_cuentas_crud",
    "description": "API CRUD para la gestion de ordenes de pago y relaciones de autorizacion.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "ImputacionPresupuestalDto": {
        "type": "object",
        "properties": {
          "Codigo": {
            "type": "string"
          },
          "Disponibilidad": {
            "type": "string"
          },
          "Nombre": {
            "type": "string"
          },
          "Registro": {
            "type": "number"
          },
          "Valor": {
            "type": "number"
          }
        },
        "required": [
          "Codigo",
          "Disponibilidad",
          "Nombre",
          "Registro",
          "Valor"
        ]
      },
      "ImpuestosRetencionesDto": {
        "type": "object",
        "properties": {
          "Base": {
            "type": "number"
          },
          "Codigo": {
            "type": "string"
          },
          "Descuento": {
            "type": "string"
          },
          "Nombre": {
            "type": "string"
          },
          "Valor": {
            "type": "number"
          },
          "CuentaContableId": {
            "type": "string"
          },
          "CuentaContableNaturaleza": {
            "type": "string"
          }
        },
        "required": [
          "Base",
          "Codigo",
          "Descuento",
          "Nombre",
          "Valor",
          "CuentaContableId",
          "CuentaContableNaturaleza"
        ]
      },
      "MovimientoContableDto": {
        "type": "object",
        "properties": {
          "Codigo": {
            "type": "string"
          },
          "Tercero": {
            "type": "string"
          },
          "NombreMovimiento": {
            "type": "string"
          },
          "Detalle": {
            "type": "string"
          },
          "Nombre": {
            "type": "string"
          },
          "Debito": {
            "type": "number"
          },
          "Credito": {
            "type": "number"
          },
          "Naturaleza": {
            "type": "string"
          },
          "CuentaContableId": {
            "type": "string"
          }
        },
        "required": [
          "Codigo",
          "Tercero",
          "NombreMovimiento",
          "Detalle",
          "Nombre",
          "Debito",
          "Credito",
          "Naturaleza",
          "CuentaContableId"
        ]
      },
      "OrdenPagoDto": {
        "type": "object",
        "properties": {
          "Fecha_creacion": {
            "format": "date-time",
            "type": "string"
          },
          "Fecha_modificacion": {
            "format": "date-time",
            "type": "string",
            "default": "2022-07-22T01:50:09.958Z"
          },
          "Activo": {
            "type": "boolean",
            "default": true
          },
          "AreaFuncional": {
            "type": "number"
          },
          "Consecutivo": {
            "type": "string"
          },
          "SolicitudGiro": {
            "type": "string"
          },
          "Vigencia": {
            "type": "number"
          },
          "DocumentoBeneficiario": {
            "type": "string"
          },
          "NombreBeneficiario": {
            "type": "string"
          },
          "Compromiso": {
            "type": "number"
          },
          "NumeroCompromiso": {
            "type": "number"
          },
          "Supervisor": {
            "type": "string"
          },
          "Detalle": {
            "type": "string"
          },
          "TipoConvenio": {
            "type": "string"
          },
          "Convenio": {
            "type": "string"
          },
          "TipoOrdenPago": {
            "type": "number"
          },
          "ActaRecibido": {
            "type": "string"
          },
          "Endoso": {
            "type": "boolean"
          },
          "BeneficiarioEndoso": {
            "type": "string"
          },
          "ValorEndoso": {
            "type": "number"
          },
          "CuentaEndoso": {
            "type": "string"
          },
          "CuentaEndosoId": {
            "type": "string"
          },
          "CuentaEndosoNaturaleza": {
            "type": "string"
          },
          "ValorOP": {
            "type": "number"
          },
          "Estado": {
            "type": "string"
          },
          "ImputacionPresupuestal": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImputacionPresupuestalDto"
            }
          },
          "Concepto": {
            "type": "string"
          },
          "ImpuestosRetenciones": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImpuestosRetencionesDto"
            }
          },
          "CuentaValorNeto": {
            "type": "string"
          },
          "CuentaValorNetoId": {
            "type": "string"
          },
          "CuentaValorNetoNaturaleza": {
            "type": "string"
          },
          "MovimientoContable": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MovimientoContableDto"
            }
          }
        },
        "required": [
          "AreaFuncional",
          "Consecutivo",
          "SolicitudGiro",
          "Vigencia",
          "DocumentoBeneficiario",
          "NombreBeneficiario",
          "Compromiso",
          "NumeroCompromiso",
          "Supervisor",
          "Detalle",
          "TipoConvenio",
          "Convenio",
          "TipoOrdenPago",
          "ActaRecibido",
          "Endoso",
          "BeneficiarioEndoso",
          "ValorEndoso",
          "CuentaEndoso",
          "CuentaEndosoId",
          "CuentaEndosoNaturaleza",
          "ValorOP",
          "Estado",
          "ImputacionPresupuestal",
          "Concepto",
          "ImpuestosRetenciones",
          "CuentaValorNeto",
          "CuentaValorNetoId",
          "CuentaValorNetoNaturaleza",
          "MovimientoContable"
        ]
      },
...
      "OrdenesPagoDto": {
        "type": "object",
        "properties": {
          "Consecutivo": {
            "type": "string"
          }
        },
        "required": [
          "Consecutivo"
        ]
      },
      "MovimientoContableDevTriDto": {
        "type": "object",
        "properties": {
          "Secuencia": {
            "type": "number"
          },
          "Codigo": {
            "type": "string"
          },
          "Nombre": {
            "type": "string"
          },
          "PorcentajeRetencion": {
            "type": "number"
          },
          "BaseRetencion": {
            "type": "number"
          },
          "Naturaleza": {
            "type": "string"
          },
          "Credito": {
            "type": "number",
            "default": 0
          },
          "Debito": {
            "type": "number",
            "default": 0
          },
          "Tercero": {
            "type": "string"
          }
        },
        "required": [
          "Secuencia",
          "Codigo",
          "Nombre",
          "PorcentajeRetencion",
          "BaseRetencion",
          "Naturaleza",
          "Credito",
          "Debito",
          "Tercero"
        ]
      },
      "DevolucionTributariaDto": {
        "type": "object",
        "properties": {
          "Fecha_creacion": {
            "format": "date-time",
            "type": "string"
          },
          "Fecha_modificacion": {
            "format": "date-time",
            "type": "string",
            "default": "2022-07-22T01:50:09.989Z"
          },
          "Activo": {
            "type": "boolean",
            "default": true
          },
          "AreaFuncional": {
            "type": "number"
          },
          "Consecutivo": {
            "type": "string"
          },
          "NumeroRequerimiento": {
            "type": "number"
          },
          "FechaSolicitud": {
            "format": "date-time",
            "type": "string"
          },
          "DocumentoBeneficiario": {
            "type": "string"
          },
          "NombreBeneficiario": {
            "type": "string"
          },
          "Concepto": {
            "type": "string"
          },
          "RazonDevolucion": {
            "type": "string"
          },
          "TipoComprobante": {
            "type": "string"
          },
          "NumeroComprobante": {
            "type": "string"
          },
          "ValorDevolucion": {
            "type": "number"
          },
          "Estado": {
            "type": "string"
          },
          "OrdenesPago": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrdenesPagoDto"
            }
          },
          "MovimientoContable": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MovimientoContableDevTriDto"
            }
          }
        },
        "required": [
          "AreaFuncional",
          "Consecutivo",
          "NumeroRequerimiento",
          "FechaSolicitud",
          "DocumentoBeneficiario",
          "NombreBeneficiario",
          "Concepto",
          "RazonDevolucion",
          "TipoComprobante",
          "NumeroComprobante",
          "ValorDevolucion",
          "Estado",
          "OrdenesPago",
          "MovimientoContable"
        ]
      },
...
    }
  },
  "paths": {
...
    "/orden-pago": {
      "post": {
        "operationId": "OrdenPagoController_post",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrdenPagoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "orden-pago"
        ]
      },
      "get": {
        "operationId": "OrdenPagoController_get",
        "parameters": [
          {
            "name": "query",
            "required": false,
            "in": "query",
            "description": "query - Filter. e.g. col1:v1,col2:v2 …",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fields",
            "required": false,
            "in": "query",
            "description": "fields - Fields returned. e.g. col1,col2 …",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortby",
            "required": false,
            "in": "query",
            "description": "sortby - Sorted-by fields. e.g. col1,col2 …",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "order - Order corresponding to each sortby field, if single value, apply to all sortby fields. e.g. desc,asc …",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "limit - Limit the size of result set. Must be an integer",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "offset - Start position of result set. Must be an integer",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "populate",
            "required": false,
            "in": "query",
            "description": "populate - show subqueries. Must be a boolean",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "orden-pago"
        ]
      }
    },
    "/orden-pago/{id}": {
      "get": {
        "operationId": "OrdenPagoController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "orden-pago"
        ]
      },
      "put": {
        "operationId": "OrdenPagoController_put",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrdenPagoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "orden-pago"
        ]
      },
      "delete": {
        "operationId": "OrdenPagoController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "orden-pago"
        ]
      }
    },
...
    "/devolucion-tributaria": {
      "post": {
        "operationId": "DevolucionTributariaController_post",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevolucionTributariaDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "devolucion-tributaria"
        ]
      },
      "get": {
        "operationId": "DevolucionTributariaController_get",
        "parameters": [
          {
            "name": "query",
            "required": false,
            "in": "query",
            "description": "query - Filter. e.g. col1:v1,col2:v2 …",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fields",
            "required": false,
            "in": "query",
            "description": "fields - Fields returned. e.g. col1,col2 …",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortby",
            "required": false,
            "in": "query",
            "description": "sortby - Sorted-by fields. e.g. col1,col2 …",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "order - Order corresponding to each sortby field, if single value, apply to all sortby fields. e.g. desc,asc …",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "limit - Limit the size of result set. Must be an integer",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "offset - Start position of result set. Must be an integer",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "populate",
            "required": false,
            "in": "query",
            "description": "populate - show subqueries. Must be a boolean",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "devolucion-tributaria"
        ]
      }
    },
    "/devolucion-tributaria/{id}": {
      "get": {
        "operationId": "DevolucionTributariaController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "devolucion-tributaria"
        ]
      },
      "put": {
        "operationId": "DevolucionTributariaController_put",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevolucionTributariaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "devolucion-tributaria"
        ]
      },
      "delete": {
        "operationId": "DevolucionTributariaController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "devolucion-tributaria"
        ]
      }
    },
...
  }
}